ci(bench): migrate benchmark tool benny to vitest
