+++ package.json
@@ -38,6 +38,7 @@
     "prepublishOnly": "npm run test && npm run build",
     "test": "jest",
     "clear-test": "jest --clearCache",
+    "bench": "vitest bench",
     "fmt": "prettier ./src/** ./tests/** -w",
     "check": "tsc --strict --noEmit --extendedDiagnostics",
     "perf": "tsc --project tests/tsconfig.json --noEmit --extendedDiagnostics",
@@ -70,11 +71,13 @@
   "homepage": "https://github.com/gvergnaud/ts-pattern#readme",
   "devDependencies": {
     "@types/jest": "^29.5.2",
+    "@vitest/coverage-istanbul": "^2.1.4",
     "jest": "^29.7.0",
     "microbundle": "^0.15.1",
     "prettier": "^2.8.8",
     "rimraf": "^5.0.1",
     "ts-jest": "^29.1.2",
-    "typescript": "^5.3.3"
+    "typescript": "^5.3.3",
+    "vitest": "^2.1.4"
   }
 }
